requirements.txt
Django>=4.2
psycopg2-binary
gunicorn
django-unfold
redis>=7.1.0